Move XBRL model to model package
